backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
djangorestframework-simplejwt
django-filter
psycopg[binary]==3.2.*
icalendar
prometheus-client
